+++ addon/globalPlugins/trainingKeyboardCommands/__init__.py
@@ -76,7 +76,7 @@ def scrapRequiredData():
 	allCommands= set()
 	tables= soup.find_all('table')
 
-	for table in tables[0:21]+tables[25: 44]:
+	for table in tables[0:22]+tables[26: 45]:
 		rows= table.find_all('tr')[1:]
 		# The following two lines for debugging purposes.
 		#tableStartsWith= rows[0].find_all('td')[0].text